Tests review + contract deployment ok

+++ code/contracts/Niel42.sol
@@ -20,13 +20,13 @@ contract Niel42 {
     // Decimals of the token
     uint8 private constant _decimals = 18;
     // Total supply of the token
-    uint96 private constant _totalSupply = uint96(1_000_000 * 10 ** _decimals);
+    uint256 private constant _totalSupply = uint256(1_000_000 * 10 ** _decimals);
 
     // Mapping from token owner address to their balance.
-    mapping(address => uint96) private _balances;
+    mapping(address => uint256) private _balances;
     // Imbricated mapping from token owner address to spender address to the
     // amount of tokens they are allowed to transfer.
-    mapping(address => mapping(address => uint96)) private _allowances;
+    mapping(address => mapping(address => uint256)) private _allowances;
     // Admins
     mapping(address => bool) private _admins;
 
@@ -38,13 +38,13 @@ contract Niel42 {
     A token contract which creates new tokens SHOULD trigger a Transfer event
     with the _from address set to 0x0 when tokens are created.
     */
-    event Transfer(address indexed _from, address indexed _to, uint96 _value);
+    event Transfer(address indexed _from, address indexed _to, uint256 _value);
 
     /*
     Approval event
-    MUST trigger on any successful call to approve(address _spender, uint96 _value).
+    MUST trigger on any successful call to approve(address _spender, uint256 _value).
     */
-    event Approval(address indexed _owner, address indexed _spender, uint96 _value);
+    event Approval(address indexed _owner, address indexed _spender, uint256 _value);
 
     /*
     Contract constructor
@@ -87,21 +87,21 @@ contract Niel42 {
     Returns the total token supply.
     The total token supply is the number of tokens available to the public.
     */
-    function totalSupply() public pure returns (uint96) {
+    function totalSupply() public pure returns (uint256) {
         return _totalSupply;
     }
 
     /*
     balanceOf function returns the balance of the given address.
     */
-    function balanceOf(address owner_) external view returns (uint96) {
+    function balanceOf(address owner_) external view returns (uint256) {
         return _balances[owner_];
     }
 
     /*
     Transfer function transfers tokens from the sender's address to the recipient's address.
     */
-    function transfer(address to_, uint96 value_) public returns (bool success) {
+    function transfer(address to_, uint256 value_) public returns (bool success) {
         require(to_ != address(0), "ERC20: transfer to the zero address");
         require(_balances[msg.sender] >= value_, "ERC20: transfer amount exceeds balance");
         _balances[msg.sender] -= value_;
@@ -114,7 +114,7 @@ contract Niel42 {
     allowance
     Returns the amount which spender_ is still allowed to withdraw from owner_.
     */
-    function allowance(address owner_, address spender_) public view returns (uint96) {
+    function allowance(address owner_, address spender_) public view returns (uint256) {
         return _allowances[owner_][spender_];
     }
 
@@ -129,7 +129,7 @@ contract Niel42 {
     for the same spender. THOUGH The contract itself shouldn’t enforce it,
     to allow backwards compatibility with contracts deployed before
     */
-    function approve(address spender_, uint96 value_) public returns (bool success) {
+    function approve(address spender_, uint256 value_) public returns (bool success) {
         require(spender_ != address(0), "ERC20: approve to the zero address");
         // require(_allowances[msg.sender][spender_] == 0 || value_ == 0,
         //         "ERC20: must first set allowance to zero");
@@ -142,17 +142,17 @@ contract Niel42 {
     Pour eviter deux calls à approve, on peut utiliser les fonctions
     increaseAllowance et decreaseAllowance
     */
-    function increaseAllowance(address spender_, uint96 addedValue) public returns (bool) {
-        uint96 newAllowance = _allowances[msg.sender][spender_] + addedValue;
+    function increaseAllowance(address spender_, uint256 addedValue) public returns (bool) {
+        uint256 newAllowance = _allowances[msg.sender][spender_] + addedValue;
         _allowances[msg.sender][spender_] = newAllowance;
         emit Approval(msg.sender, spender_, newAllowance);
         return true;
     }
 
-    function decreaseAllowance(address spender_, uint96 subtractedValue) public returns (bool) {
-        uint96 currentAllowance = _allowances[msg.sender][spender_];
+    function decreaseAllowance(address spender_, uint256 subtractedValue) public returns (bool) {
+        uint256 currentAllowance = _allowances[msg.sender][spender_];
         require(currentAllowance >= subtractedValue, "ERC20: decreased allowance below zero");
-        uint96 newAllowance = currentAllowance - subtractedValue;
+        uint256 newAllowance = currentAllowance - subtractedValue;
         _allowances[msg.sender][spender_] = newAllowance;
         emit Approval(msg.sender, spender_, newAllowance);
         return true;
@@ -169,7 +169,7 @@ contract Niel42 {
     authorized the sender of the message via some mechanism.
     Note Transfers of 0 values MUST be treated as normal transfers and fire the Transfer event.
     */
-    function transferFrom(address from_, address to_, uint96 value_) public returns (bool success) {
+    function transferFrom(address from_, address to_, uint256 value_) public returns (bool success) {
         require(value_ <= _balances[from_], "ERC20: transfer amount exceeds balance");
         require(value_ <= _allowances[from_][msg.sender], "ERC20: insufficient allowance");
         require(to_ != address(0), "ERC20: transfer to the zero address");
